% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTrace.R
\name{geotrace}
\alias{geotrace}
\alias{print.geotrace}
\alias{plot.geotrace}
\title{geotrace:}
\usage{
geotrace(case,origins,pop, nsim=999, prob = NULL, agg="median",
      dist_type = ifelse(isTRUE(st_is_longlat(case)), "Great Circle", "Euclidean"),
      alternative = "two.sided", orig_names = NULL, seed = NULL)
\method{print}{geotrace}(x, ...)
\method{plot}{geotrace}(x, ...)
}
\arguments{
\item{case}{sf object, cases data}

\item{origins}{sf object, potential origins of infectious diseases}

\item{pop}{sf object, population of residences}

\item{nsim}{int, repeat times of sampling}

\item{prob}{vector of probability weights for obtaining the elements of the
vector being sampled}

\item{agg}{character, distance aggregation method, either (default), or "mean".}

\item{dist_type}{character; for Cartesian coordinates only: one of Euclidean,
Hausdorff or Frechet; for geodetic coordinates, great circle
distances are computed; sf::st_distance()}

\item{alternative}{character, the alternative hypothesis: either "two.sided" (default),
"greater" or "less".}

\item{orig_names}{a vector of the names of potential origins}

\item{seed}{int, random seed}

\item{x}{a list of \code{geotrace} result}

\item{...}{Ignore}
}
\description{
Analysis potential origins of infectious diseases from spatial distance
}
\examples{
###############
## JohnSnow 1854 Broad Street cholera outbreak
###############
cases <- house_death[house_death$death_dum==1,]
BS_cholera <- geotrace(cases, pumps, house_death)
BS_cholera
plot(BS_cholera)

\dontrun{
#############
## Origins of Wuhan Covid-19
#############
orig_names <- location$Name
WH_cov19 <- geotrace(covid19_wh,location, wuhan_pop, prob = wuhan_pop$Wuhan_PPP,
           orig_names = orig_names)
WH_cov19
plot(WH_cov19)
}
}
